Health and Fitness platform for student in Walter Sisulu University
Given the task of solving a problem in the community, MMEA is a group of students
who have taken on the responsibility of promoting Health and Fitness in the Eastern
Cape.  The MMEA team offers a variety of services to address the lack of Health and
Fitness while also promoting Overall Wellness.

MMEA has partnered with small businesses in East London to provide them with People, a Digital Platform, and an Online Presence. It conducted interviews with Health Specialists, Nutritionists, and Sports Coaches to learn more about Health & Fitness. Furthermore, the team of developers started sourcing fresh produce in the Amathole district farms in the Eastern Cape and making it available to students via a
digital platform.

As technology continues to advance, information’s disseminates easily with no
hassles, and it has been a common part of our everyday needs in this world.

Rapid advances in technology have helped to meet these demands of society.